Update for Ubuntu 24.04
Switch the Dockerfile to using
the ubuntu user/group which is
required for building after
https://github.com/davidfrantz/base_image/pull/4
has been merged.

Also remove the WORKDIR since that
is the last thing set by the Dockerfile
for davidfrantz/base.

Include paths for some dependencies
have changed, so update the Makefile
with those paths as well.