Fail `force-cube` explicitly for missing features
Previously, an vector file (e.g. GPKG) could be supplied with
a corresponding layer name but a feature count of 0. This leads
to `force-cube` getting stuck.

As this was discovered when using `force-tile-extent`, a check for
the return value was added there with an error message indicating
to the user what command to re-run to get more information during
which step `force-cube` failed